+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Optional inline SVG workaround ([#1](https://github.com/marp-team/marpit/pull/1))
 * Split the injection of markdown-it plugins and provide interface of markdown-it plugin ([#2](https://github.com/marp-team/marpit/pull/2))
+* Split rendering style into Marpit#renderStyle ([#3](https://github.com/marp-team/marpit/pull/3))
 * Add JSDoc about `Marpit` class
 
 ## v0.0.0 - 2018-03-24